Half-star ratings, titles on the shelf, and a shelf menu in the detail sheet
- Ratings move from Int to Double in 0.5 steps. StarRatingView draws
  half stars and StarRatingPicker snaps a tap or drag to the nearest
  half. Whole-number ratings written earlier decode unchanged, and the
  Supabase schema's rating column widens to match.
- The profile shelf shows each book's title between the cover and its
  stars, on a fixed two-line block so star rows stay aligned across a
  grid row. A missing jacket now falls back to a placeholder mark rather
  than repeating the title.
- Drop the "Good read" / "A new favorite" text under the stars, in both
  the review sheet and the shelf detail.
- BookDetailView gets the shelf menu: Want to Read / Reading / Read with
  a checkmark on the current shelf, plus Remove from Library once the
  book is on one. It reads and writes LibraryStore directly, so it works
  for a book that isn't in the library yet.
- The local fallback user id moves to AuthService.effectiveUserId
  instead of being duplicated per call site.

+++ BookApp/Services/SupabaseService.swift
@@ -137,7 +137,7 @@ final class SupabaseService {
         return try decoder.decode([Review].self, from: data)
     }
 
-    func upsertReview(userId: UUID, googleBooksId: String, rating: Int, reviewText: String?) async throws -> Review {
+    func upsertReview(userId: UUID, googleBooksId: String, rating: Double, reviewText: String?) async throws -> Review {
         let body: [String: Any?] = [
             "user_id": userId.uuidString,
             "google_books_id": googleBooksId,
